Add files via upload
All files except Boolean need to be tested. Rotary, LED, and RGB need to be rewritten in C++ to be imported into Arduino IDE. Boolean library is written and tested. Can be further optimized. https://docs.arduino.cc/learn/contributions/arduino-creating-library-guide. This is the link on how to create Arduino libraries.